// src/interfaces/general.ts
export interface ISeoData {
  title: string;
  description: string;
  keywords: string;
  author: string;
  ogTitle: string;
  ogDescription: string;
  ogUrl: string;
  ogImage: string;
  twitterTitle: string;
  twitterDescription: string;
  twitterImage: string;
  favicon: string;
}

export interface IGeneral {
  company_name: string;
  logo: string;
  description: string;
}

export interface IOffer {
  title: string;
  description: string;
}